pyproject.toml
[project]
name = "automated-followup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "google-generativeai>=0.8.5",
    "sqlglot>=27.28.1",
]
